% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate variables in a file}
\usage{
interpolate(filename, values, sentinel = "\%")
}
\arguments{
\item{filename}{the name of the file containing the text to be interpolated}

\item{values}{a named vector, in which the names are used as the variable names, and the 
values as the variable values.}

\item{sentinel}{a sentinel character indicating a variable in the text. Defaults to "\%".}
}
\description{
This function reads the contents of a file and performs a search-and-replace
on a set of given variable names with given values. In the context of a CGI script,
this function is useful for interpolating HTML templates and using them as the HTTP response.
variable occurrences are marked with a sentinel character, which defaults to "\%".
}
\details{
For example, given the text "Hello \%foo" and the variable foo="World", the interpolation
will produce the text "Hello World"
}
\examples{
error <- "Insufficient amount of coffee!"
errorHTML <- interpolate("../../html/app/error.html",c(message=error))
respondHTML(errorHTML)
}
