% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{readGET}
\alias{readGET}
\title{Read GET data from HTTP request}
\usage{
readGET()
}
\value{
a named list containing the GET data
}
\description{
Reads the GET data transmitted via an HTTP request and returns it as a named character vector.
}
\examples{
getData <- readGET()
if ("foo" \%in\% names(getData)) {
   foo <- getData[["foo"]]
}
}
