% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{readPOST}
\alias{readPOST}
\title{Read POST data from HTTP request}
\usage{
readPOST()
}
\value{
a named list containing the POST data or NULL if none exists
}
\description{
Reads the POST data transmitted via an HTTP request and returns it as a named character vector.
}
\examples{
postData <- readPOST()
if ("foobar" \%in\% names(postData)) {
   foobar <- postData[["foobar"]]
}
}
