% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respond}
\alias{respond}
\title{Respond to HTTP request using a given MIME type}
\usage{
respond(content, mime)
}
\arguments{
\item{content}{a character string containing the content}

\item{mime}{a character string containing the MIME type (e.g. "text/html" or "image/gif")}
}
\description{
Responds to an HTTP request using a given MIME type
}
\examples{
error <- "Insufficient amount of coffee!"
errorHTML <- interpolate("../../html/app/error.html",c(message=error))
respondHTML(errorHTML)
}
