% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respond400}
\alias{respond400}
\title{Respond to HTTP request with a 400:BadRequest error}
\usage{
respond400(message)
}
\arguments{
\item{message}{the error message}
}
\description{
Responds to an HTTP request with 400:BadRequest error
}
