% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respond404}
\alias{respond404}
\title{Respond to HTTP request with a 404:NotFound error}
\usage{
respond404(message)
}
\arguments{
\item{message}{the error message}
}
\description{
Responds to an HTTP request with 404:NotFound error
}
