% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respond500}
\alias{respond500}
\title{Respond to HTTP request with a 500 Internal Server Error}
\usage{
respond500(message)
}
\arguments{
\item{message}{the error message}
}
\description{
Responds to an HTTP request with 500 Internal Server Error
}
