% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondBinary}
\alias{respondBinary}
\title{Respond to HTTP request using binary data in a given MIME type}
\usage{
respondBinary(binFile, mime)
}
\arguments{
\item{binFile}{the name of a file containing the binary data}

\item{mime}{a character string containing the MIME type (e.g. "image/gif")}
}
\description{
Responds to an HTTP request using binary data a given MIME type
}
\examples{
error <- "Insufficient amount of coffee!"
errorHTML <- interpolate("../../html/app/error.html",c(message=error))
respondHTML(errorHTML)
}
