% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondHTML}
\alias{respondHTML}
\title{Respond to HTTP request using HTML}
\usage{
respondHTML(html)
}
\arguments{
\item{html}{a character string containing the HTML content}
}
\description{
Responds to an HTTP request using a given HTML content
}
\examples{
error <- "Insufficient amount of coffee!"
errorHTML <- interpolate("../../html/app/error.html",c(message=error))
respondHTML(errorHTML)
}
