% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondJSON}
\alias{respondJSON}
\title{Respond to HTTP request using JSON}
\usage{
respondJSON(content)
}
\arguments{
\item{content}{a named list, which will be serialized to JSON using RJSONIO}
}
\description{
Responds to an HTTP request using JSON formatted data
}
\examples{
data <- list(foo="bar",baz=1:5)
respondJSON(data)
}
