% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondSVG}
\alias{respondSVG}
\title{Respond to HTTP request with SVG data}
\usage{
respondSVG(svgFile, filename = "image.svg", download = FALSE)
}
\arguments{
\item{svgFile}{the name of a file containing the binary data}

\item{filename}{the filename to be used in the HTTP header. Defaults to image.svg}

\item{download}{whether or not to treat this as a download. Defaults to FALSE.}
}
\description{
Responds to an HTTP request with SVG data from a given SVG file
}
