% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondTemplateHTML}
\alias{respondTemplateHTML}
\title{Respond to HTTP request using HTML interpolated from a template file}
\usage{
respondTemplateHTML(templateFile, values)
}
\arguments{
\item{templateFile}{the name of the file containing the HTML content}

\item{values}{a named vector, in which the names are used as the variable names, and the 
values as the variable values.}
}
\description{
Responds to an HTTP request using a given HTML file as well as interpolated content
}
\examples{
error <- "Insufficient amount of coffee!"
respondTemplateHTML("../../html/app/error.html",c(message=error))
}
