% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondTEXT}
\alias{respondTEXT}
\title{Respond to HTTP request using plain text}
\usage{
respondTEXT(text)
}
\arguments{
\item{content}{a character string to respond}
}
\description{
Responds to an HTTP request using plain text
}
\examples{
respondTEXT("Hello World!")
}
