% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondBinary}
\alias{respondBinary}
\title{Respond to HTTP request using binary data in a given MIME type}
\usage{
respondBinary(binFile, mime, filename, download = FALSE)
}
\arguments{
\item{binFile}{the name of a file containing the binary data}

\item{mime}{a character string containing the MIME type (e.g. "image/gif")}

\item{filename}{the filename to be used in the HTTP header}

\item{download}{logical. whether or not to mark this as a download}
}
\description{
Responds to an HTTP request using binary data a given MIME type
}
\examples{
respondBinary(imageFile,"image/png","image.png",download=TRUE)
}
