% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondPDF}
\alias{respondPDF}
\title{Respond to HTTP request with PDF data}
\usage{
respondPDF(pdfFile)
}
\arguments{
\item{pdfFile}{the name of a file containing the binary data}
}
\description{
Responds to an HTTP request with PDF data from a given PDF file
}
