% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{respondPNG}
\alias{respondPNG}
\title{Respond to HTTP request with PNG image data}
\usage{
respondPNG(pngFile)
}
\arguments{
\item{pngFile}{the name of a file containing the binary data}
}
\description{
Responds to an HTTP request with PNG image data from a given PNG file
}
