% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgir.R
\name{setMessageSink}
\alias{setMessageSink}
\title{Redirect messages to a log file}
\usage{
setMessageSink(filename = "msg.log")
}
\description{
Redirects any messages (e.g. errors) to a log file. Doing so at the start of your script is
absolutely vital! Otherwise any potential info messages, warnings or errors can corrupt your
HTTP response
}
